+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+### Fixed
+
+- Fix admin order edit tool to use correct id for supplier query
+
 ## [1.11.5] - 2020-07-07
 
 ### Added